deployment_package_2/package/requirements.txt
anthropic
google-generativeai